--- core/domain/src/main/java/com/zcard/domain/usecase/InitCardEditorUseCase.kt
@@ -1,5 +1,0 @@
-package com.zcard.domain.usecase
-
-interface InitCardEditorUseCase {
-    suspend operator fun invoke(cardId: Long): Result<LoadCardEditorResult>
-}
